// src/dishhive-web/src/app/api/generated/dishhive-api.client.ts
export enum FirstDayOfWeek {
    Monday = "Monday",
    Sunday = "Sunday",
}
// ...
export enum MeasurementSystem {
    Metric = "Metric",
    Imperial = "Imperial",
}
// ...
export class SupportedLanguageDto implements ISupportedLanguageDto {
    code!: string;
    displayName!: string;

    [key: string]: any;

    constructor(data?: ISupportedLanguageDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.code = _data["code"];
            this.displayName = _data["displayName"];
        }
    }

    static fromJS(data: any): SupportedLanguageDto {
        data = typeof data === 'object' ? data : {};
        let result = new SupportedLanguageDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["code"] = this.code;
        data["displayName"] = this.displayName;
        return data;
    }
}

export interface ISupportedLanguageDto {
    code: string;
    displayName: string;

    [key: string]: any;
}
// ...
export class UpdateUserPreferencesDto implements IUpdateUserPreferencesDto {
    measurementSystem?: MeasurementSystem | undefined;
    firstDayOfWeek?: FirstDayOfWeek | undefined;
    preferredLanguage?: string | undefined;
    translateImportedRecipes?: boolean | undefined;

    [key: string]: any;

    constructor(data?: IUpdateUserPreferencesDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.measurementSystem = _data["measurementSystem"];
            this.firstDayOfWeek = _data["firstDayOfWeek"];
            this.preferredLanguage = _data["preferredLanguage"];
            this.translateImportedRecipes = _data["translateImportedRecipes"];
        }
    }

    static fromJS(data: any): UpdateUserPreferencesDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateUserPreferencesDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["measurementSystem"] = this.measurementSystem;
        data["firstDayOfWeek"] = this.firstDayOfWeek;
        data["preferredLanguage"] = this.preferredLanguage;
        data["translateImportedRecipes"] = this.translateImportedRecipes;
        return data;
    }
}

export interface IUpdateUserPreferencesDto {
    measurementSystem?: MeasurementSystem | undefined;
    firstDayOfWeek?: FirstDayOfWeek | undefined;
    preferredLanguage?: string | undefined;
    translateImportedRecipes?: boolean | undefined;

    [key: string]: any;
}
// ...
export class UserPreferencesDto implements IUserPreferencesDto {
    measurementSystem!: MeasurementSystem;
    firstDayOfWeek!: FirstDayOfWeek;
    preferredLanguage!: string;
    translateImportedRecipes!: boolean;
    supportedLanguages!: SupportedLanguageDto[];

    [key: string]: any;

    constructor(data?: IUserPreferencesDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.supportedLanguages = [];
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.measurementSystem = _data["measurementSystem"];
            this.firstDayOfWeek = _data["firstDayOfWeek"];
            this.preferredLanguage = _data["preferredLanguage"];
            this.translateImportedRecipes = _data["translateImportedRecipes"];
            if (Array.isArray(_data["supportedLanguages"])) {
                this.supportedLanguages = [] as any;
                for (let item of _data["supportedLanguages"])
                    this.supportedLanguages!.push(SupportedLanguageDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): UserPreferencesDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserPreferencesDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["measurementSystem"] = this.measurementSystem;
        data["firstDayOfWeek"] = this.firstDayOfWeek;
        data["preferredLanguage"] = this.preferredLanguage;
        data["translateImportedRecipes"] = this.translateImportedRecipes;
        if (Array.isArray(this.supportedLanguages)) {
            data["supportedLanguages"] = [];
            for (let item of this.supportedLanguages)
                data["supportedLanguages"].push(item ? item.toJSON() : undefined as any);
        }
        return data;
    }
}

export interface IUserPreferencesDto {
    measurementSystem: MeasurementSystem;
    firstDayOfWeek: FirstDayOfWeek;
    preferredLanguage: string;
    translateImportedRecipes: boolean;
    supportedLanguages: SupportedLanguageDto[];

    [key: string]: any;
}